refactor(svelte): replace store-based architecture with native Svelte 5 runes
- Replace writable/derived/fromStore pattern with $state/$derived.by/$effect
- Extract useStreamLGP into stream.svelte.ts using native runes
- Convert stream.custom.ts to stream.custom.svelte.ts with runes
- Add svelte-module build plugin to compile .svelte.ts files
- Add ambient rune type declarations for TypeScript support
- Consolidate duplicate context APIs: deprecate getStream() in favor
  of getStreamContext() as the single getter
- Keep setStreamContext() and provideStream() as complementary setters
- Update tests, README, and ESLint config
